Replace long long with int64_t in some places

+++ wvm/cpp/src/Util.cpp
@@ -28,13 +28,13 @@ namespace WVM::Util {
 		return out;
 	}
 
-	bool parseUL(const std::string &str, unsigned long long &out, int base) {
+	bool parseUL(const std::string &str, uint64_t &out, int base) {
 		char *endptr;
 		out = strtoull(str.c_str(), &endptr, base);
 		return static_cast<unsigned long>(endptr - str.c_str()) == str.size();
 	}
 
-	bool parseLong(const std::string &str, long long &out, int base) {
+	bool parseLong(const std::string &str, int64_t &out, int base) {
 		char *endptr;
 		out = strtoll(str.c_str(), &endptr, base);
 		return static_cast<unsigned long>(endptr - str.c_str()) == str.size();
